Fase 3/requirements.txt

faiss-cpu==1.7.4
sentence-transformers==2.7.0
ollama==0.1.8
streamlit==1.36.0
tomli==2.0.1
numpy==1.26.4
huggingface-hub==0.35.1
transformers==4.56.2